Moved Flat File Autoconfigs, props and tests to the flatfile package
