+++ spring-cloud-starter-single-step-batch-job/src/main/java/org/springframework/cloud/task/batch/autoconfigure/flatfile/FlatFileItemWriterAutoConfiguration.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.task.batch.autoconfigure;
+package org.springframework.cloud.task.batch.autoconfigure.flatfile;
 
 import java.util.ArrayList;
 import java.util.List;
